Application target thanks to George Ferguson <george@georgeferguson.com>.
His comment:

I've used your keyboardViewer program since Snow Leopard screwed up my old ways of getting an on-screen keyboard (for a touchscreen I have on an old mini in my kitchen). But I needed way to launch it from said touchscreen. Rather than write a script to run the executable and use the Finder to put an icon on that, I added a target quickly to your xcode project and let it build the app bundle.